+++ src/app/search/search-filters.ts
@@ -245,6 +245,7 @@ export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
       reacquirable: ['reacquirable'],
       hasLight: ['light', 'haslight', 'haspower'],
       complete: ['goldborder', 'yellowborder', 'complete'],
+      curated: ['curated'],
       wishlist: ['wishlist'],
       wishlistdupe: ['wishlistdupe'],
       masterwork: ['masterwork', 'masterworks'],
@@ -840,21 +841,12 @@ function searchFilters(
           : item.owner === stores[storeIndex].id;
       },
       classType(item: DimItem, predicate: string) {
-        let value;
-
-        switch (predicate) {
-          case 'titan':
-            value = 0;
-            break;
-          case 'hunter':
-            value = 1;
-            break;
-          case 'warlock':
-            value = 2;
-            break;
+        const classes = ['titan', 'hunter', 'warlock'];
+        if (item.classified) {
+          return false;
         }
 
-        return item.classType === value;
+        return item.classType === classes.indexOf(predicate);
       },
       glimmer(item: DimItem, predicate: string) {
         const boosts = [
@@ -1207,6 +1199,36 @@ function searchFilters(
       hasLight(item: DimItem) {
         return item.primStat && statHashes.has(item.primStat.statHash);
       },
+      curated(item: D2Item) {
+        if (!item) {
+          return false;
+        }
+
+        // TODO: remove if there are no false positives, as this precludes maintaining a list for curatedNonMasterwork
+        // const masterWork = item.masterworkInfo && item.masterworkInfo.statValue === 10;
+        // const curatedNonMasterwork = [792755504, 3356526253, 2034817450].includes(item.hash); // Nightshade, Wishbringer, Distant Relation
+
+        const legendaryWeapon =
+          item.bucket && item.bucket.sort === 'Weapons' && item.tier.toLowerCase() === 'legendary';
+
+        const oneSocketPerPlug =
+          item.sockets &&
+          item.sockets.sockets
+            .filter(
+              // Remove Ornaments and Trackers
+              (socket) =>
+                socket &&
+                socket.plug &&
+                ![2947756142, 3940152116].includes(socket.plug.plugItem.plug.plugCategoryHash)
+            )
+            .every((socket) => socket && socket.plugOptions.length === 1);
+
+        return (
+          legendaryWeapon &&
+          // (masterWork || curatedNonMasterwork) && // checks for masterWork(10) or on curatedNonMasterWork list
+          oneSocketPerPlug
+        );
+      },
       weapon(item: DimItem) {
         return item.bucket && item.bucket.sort === 'Weapons';
       },
@@ -1250,7 +1272,7 @@ function searchFilters(
           item.sockets.sockets.some((socket) => {
             return !!(
               socket.plug &&
-              ![2323986101, 2600899007, 1835369552, 3851138800].includes(
+              ![2323986101, 2600899007, 1835369552, 3851138800, 791435474].includes(
                 socket.plug.plugItem.hash
               ) &&
               socket.plug.plugItem.plug &&