+++ config/webpack.js
@@ -32,7 +32,7 @@ module.exports = (env) => {
     if (!fs.existsSync('key.pem') || !fs.existsSync('cert.pem')) {
       console.log('Generating certificate');
       execSync(
-        "openssl req -newkey rsa:2048 -new -nodes -x509 -days 3650 -keyout key.pem -out cert.pem -subj '/CN=www.mydom.com/O=My Company Name LTD./C=US'"
+        'openssl req -newkey rsa:2048 -new -nodes -x509 -days 3650 -keyout key.pem -out cert.pem -subj "/CN=www.mydom.com/O=My Company Name LTD./C=US"'
       );
     }
   }
@@ -152,9 +152,10 @@ module.exports = (env) => {
             {
               loader: 'css-loader',
               options: {
-                modules: true,
-                camelCase: true,
-                localIdentName: isDev ? '[name]_[local]-[hash:base64:5]' : '[hash:base64:5]',
+                modules: {
+                  localIdentName: isDev ? '[name]_[local]-[hash:base64:5]' : '[hash:base64:5]'
+                },
+                localsConvention: 'camelCaseOnly',
                 sourceMap: true
               }
             },