+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next
+* Updated is:modded to take into account for activity mods.
 
 # 5.35.0 (2019-06-30)
 * Removed is:curated as an alias for is:wishlist.