+++ docs/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Next
 * Fixed: Classified items required some finesse.
+* Updated is:modded to take into account for activity mods.
 
 # 5.35.0 (2019-06-30)
 * Removed is:curated as an alias for is:wishlist.
+* Re-added is:curated as a filter for Bungie curated rolls.
 
 # 5.34.0 (2019-06-23)