+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next
+* Fixed: Classified items required some finesse.
 * Updated is:modded to take into account for activity mods.
 
 # 5.35.0 (2019-06-30)