+++ docs/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Next
+* Added perkname: search.
 * Charge Time and Draw Time now compare correctly!
 * Fixed: Classified items required some finesse.
 * Updated is:modded to take into account for activity mods.
 * Re-added is:curated as a filter for Bungie curated rolls.
-
 
 # 5.35.0 (2019-06-30)
 * Removed is:curated as an alias for is:wishlist.